requirements.txt
aniso8601==3.0.2
Django==2.1.4
django-graphql-jwt==0.2.0
djangorestframework==3.9.0
graphene==2.1.3
graphene-django==2.2.0
graphql-core==2.1
graphql-relay==0.4.5
promise==2.2.1
PyJWT==1.7.1
pytz==2018.7
Rx==1.6.1
singledispatch==3.4.0.3
six==1.12.0